Add files via upload
Hey team. Had a really rough go with the JS, so if someone could have a look and fix it up. Was totally lost there.
The CSS also needs some polish, but simply let me know what or where I went wrong within the CSS and I can fix that on my own.